Conflicts:
	OneFlower.vcxproj.filters
	Project/Component/PickupComponent.cpp
	Project/Component/PickupComponent.hpp
	Project/GraphicalUserInterface/ButtonDefiner.cpp
	Project/GraphicalUserInterface/GraphicalInventoryDefiner.cpp
	Project/GraphicalUserInterface/GraphicalUserInterface.cpp
	Project/GraphicalUserInterface/GraphicalUserInterfaceHandler.cpp
	Project/GraphicalUserInterface/ScrollBarDefiner.cpp
	Project/GraphicalUserInterface/WindowIconDefiner.cpp
	Project/Text/FloatingTextDefiner.cpp
	Project/Text/MessageDefiner.cpp
	Project/WindowHandler.cpp

+++ Project/Text/InputBox.hpp
@@ -9,7 +9,7 @@ class InputBox:public Message
 public:
 
 	InputBox();
-	InputBox(sf::Font& font,int len,int xStart,int yStart,int xEnd,int yEnd);
+	InputBox(sf::Font& font, int len, float xStart, float yStart, float xEnd, float yEnd);
 
 	void draw();
 	void move(int x,int y);
